requirements.txt
aiomysql==0.2.0
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.2.0
distlib==0.3.6
Django==4.2.1
django-filter==23.3
et-xmlfile==1.1.0
filelock==3.12.0
idna==3.4
openpyxl==3.1.2
pika==1.3.1
Pillow==9.5.0
pipenv==2023.4.29
platformdirs==3.5.0
PyMySQL==1.1.0
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.23.0
virtualenv-clone==0.5.7
mysqlclient==2.1.1
